requirements.txt
# Voice Dictation Assistant Dependencies
# Core audio and speech recognition
pyaudio>=0.2.14  # Updated for Python 3.12 compatibility
assemblyai==0.17.0

# AI text processing
openai==1.5.0

# Windows integration and automation
pywin32==306
pyautogui==0.9.54
pygetwindow==0.0.9
pyperclip==1.8.2

# Hotkey management
global-hotkeys==0.1.7

# System tray and GUI
pystray>=0.19.5
Pillow>=10.0.0
win10toast>=0.9

# Configuration and utilities
pyyaml==6.0.1
pydantic==2.5.0

# Testing framework and dependencies
pytest==7.4.0
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-html==4.1.1
pytest-xdist==3.3.1
psutil==5.9.6

# Code quality and linting
flake8==6.1.0
black==23.11.0
isort==5.12.0
mypy==1.7.1

# Security scanning
bandit==1.7.5
safety==2.3.5

# Additional dependencies for enhanced functionality
scipy>=1.11.0
numpy>=1.24.0
requests>=2.31.0

# Note: PyAudio installation may require manual setup on Windows with Python 3.12
# Alternative: Use sounddevice or pyaudio-wheels for audio capture 